refactor: split the trace file plumbing and ffi adapter tests into sibling modules
